+++ lib/rust/parser/generate-java/java/org/enso/syntax2/UnsupportedSyntaxException.java
@@ -1,0 +1,14 @@
+package org.enso.syntax2;
+
+public final class UnsupportedSyntaxException extends Exception {
+    protected final Tree tree;
+
+    UnsupportedSyntaxException(Tree treeIn) {
+        super("Tree contains unsupported syntax. Details are in an `Unsupported` node in the tree.");
+        tree = treeIn;
+    }
+
+    public final Tree getTree() {
+                              return tree;
+                                          }
+}
